FE-Refactor: 타입 분리함
