+++ apps/web/api/items.ts
@@ -2,6 +2,10 @@ import { API_ENDPOINTS } from "@repo/constants";
 import { type ISeat } from "@repo/types";
 import { axiosRequester } from "@/lib/axios";
 
+/**
+ * 모든 좌석 데이터를 가져옵니다.
+ * @returns 모든 좌석 데이터를 포함하는 Promise.
+ */
 export const getAllSeats = async (): Promise<ISeat[]> => {
   const { data } = await axiosRequester<ISeat[]>({
     options: {
@@ -13,6 +17,10 @@ export const getAllSeats = async (): Promise<ISeat[]> => {
   return data;
 };
 
+/**
+ * 특정 아이템 데이터를 수정합니다.
+ * @returns 수정 결과 메시지를 포함하는 Promise.
+ */
 export const patchItem = async (itemId: string, formData: FormData): Promise<string> => {
   const { data } = await axiosRequester<string, FormData>({
     options: {
