+++ packages/types/src/responseType.ts
@@ -1,3 +1,4 @@
+import { IReservation } from "./reservationType";
 import { IUser } from "./userType";
 
 export interface ResponseType<T> {
@@ -11,4 +12,9 @@ export interface SignInResponseType<T> {
   message?: string;
   status?: "success" | "error";
   user?: IUser;
+}
+
+export interface ReservationResponse {
+  message: string;
+  savedReservation: IReservation[];
 }
